requirements.txt
 spade
 pyyaml
 aiohttp_cors
